refactored to typescript
