+++ package.json
@@ -4,7 +4,7 @@
   "description": "medir tus habilidades en JavaScript y ver cómo puedes resolver los desafíos propuestos",
   "main": "index.js",
   "scripts": {
-    "start": "node src/index.js"
+    "start": "node dist/index.js"
   },
   "keywords": [
     "javascript",
@@ -13,6 +13,10 @@
   "author": "Alejandro Lopez <alejandro.lopez@tiui.mx>",
   "license": "MIT",
   "dependencies": {
+    "axios": "^1.7.2",
     "xmlhttprequest": "^1.8.0"
+  },
+  "devDependencies": {
+    "typescript": "^5.5.2"
   }
 }
