+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "compilerOptions": {
+    "target": "es6",                    // or a target suitable for your Node.js version
+    "module": "commonjs",               // or "esnext" if you use ES modules
+    "outDir": "./dist",                 // output directory for compiled files
+    "strict": true,                     // optional, enables strict type-checking
+    "esModuleInterop": true,            // allows interoperability between CommonJS and ES Modules
+    "resolveJsonModule": true,          // allows importing JSON modules
+    "skipLibCheck": true,               // optional, skips type-checking of declaration files
+    "forceConsistentCasingInFileNames": true
+  },
+  "include": ["src/**/*.ts"],           // adjust path to match your source files
+  "exclude": ["node_modules"]
+}
